backend/requirements.txt
Django==5.1.*
djangorestframework==3.15.*
django-cors-headers==4.4.*
django-filter==24.*
psycopg[binary]==3.2.*
openpyxl==3.1.*
yfinance==0.2.*
apscheduler>=3.10,<4
